+++ apps/backend-functions/src/data/internals.ts
@@ -7,7 +7,7 @@ import { db } from '../internals/firebase';
 import { OrganizationDocument } from './types';
 import { PermissionsDocument } from '@blockframes/permissions/types';
 import { ContractDocument } from '@blockframes/contract/contract/+state/contract.firestore';
-import { createImgRef } from '@blockframes/utils/image-uploader';
+import { createImgRef } from '@blockframes/utils/media/media.firestore';
 import { createDenomination } from '@blockframes/organization/+state/organization.firestore';
 import { App, getOrgAppAccess, getSendgridFrom } from '@blockframes/utils/apps';
 import { appUrlMarket, appUrlContent } from '../environments/environment';
