+++ libs/contract/src/lib/contract/+state/contract.model.ts
@@ -11,7 +11,7 @@ import {
   PublicContractDocumentWithDates
 } from './contract.firestore';
 import { createParty } from '@blockframes/utils/common-interfaces/identity';
-import { createImgRef } from '@blockframes/utils/image-uploader';
+import { createImgRef } from '@blockframes/utils/media';
 import {
   ContractVersion,
   ContractVersionWithTimeStamp,
