+++ libs/event/src/lib/components/card/card.component.html
