+++ libs/movie/src/lib/movie/components/card/movie-image.pipe.ts
@@ -1,6 +1,6 @@
 import { Pipe, PipeTransform, NgModule } from '@angular/core';
 import { Movie } from '@blockframes/movie/+state/movie.model';
-import { ImgRef } from '@blockframes/utils/image-uploader';
+import { ImgRef } from '@blockframes/utils/media';
 import { CommonModule } from '@angular/common';
 
 @Pipe({
