+++ libs/movie/src/lib/movie/components/import/view-extracted-elements/movies/view-extracted-movies.component.ts
@@ -38,7 +38,7 @@ import { createStakeholder } from '@blockframes/utils/common-interfaces/identity
 import { createRange, createPrice } from '@blockframes/utils/common-interfaces';
 import { Intercom } from 'ng-intercom';
 import { cleanModel, getKeyIfExists } from '@blockframes/utils/helpers';
-import { ImageUploader } from '@blockframes/utils/image-uploader';
+import { ImageUploader } from '@blockframes/utils/media';
 import { DynamicTitleService } from '@blockframes/utils/dynamic-title/dynamic-title.service';
 import { MovieImportState } from '../../import-utils';
 import { createDistributionRight } from '@blockframes/distribution-rights/+state';
