+++ libs/notification/src/lib/notification/+state/notification.model.ts
@@ -1,5 +1,5 @@
 import { NotificationDocument } from "./notification.firestore";
-import { ImgRef } from "@blockframes/utils/image-uploader";
+import { ImgRef } from "@blockframes/utils/media";
 
 // export type Notification = NotificationDocument;
