+++ libs/ui/src/lib/media/cropper/cropper/cropper.component.ts
@@ -6,7 +6,7 @@ import { Observable, BehaviorSubject, of, Subscription, combineLatest } from 'rx
 import { zoom, zoomDelay, check, finalZoom } from '@blockframes/utils/animations/cropper-animations';
 import { AngularFireStorage, AngularFireStorageReference } from '@angular/fire/storage';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { ImgRef, createImgRef } from '@blockframes/utils/image-uploader';
+import { ImgRef, createImgRef } from '@blockframes/utils/media';
 import { sanitizeFileName } from '@blockframes/utils/file-sanitizer';
 import { delay } from '@blockframes/utils/helpers';
