+++ libs/user/src/lib/auth/forms/profile-edit.form.ts
@@ -1,6 +1,6 @@
 import { FormControl } from '@angular/forms';
 import { User } from '@blockframes/auth/+state/auth.store';
-import { createImgRef, ImgRef } from '@blockframes/utils/image-uploader';
+import { createImgRef, ImgRef } from '@blockframes/utils/media';
 import { FormEntity } from '@blockframes/utils/form/forms/entity.form';
 
 export interface Profile {
