+++ libs/user/src/lib/user/+state/user.firestore.ts
@@ -1,4 +1,4 @@
-import { ImgRef, createImgRef } from "@blockframes/utils/image-uploader";
+import { ImgRef, createImgRef } from "@blockframes/utils/media";
 
 export interface User {
   uid: string;
