+++ libs/user/src/lib/user/+state/user.model.ts
@@ -1,6 +1,6 @@
 import { UserRole } from "@blockframes/permissions/types";
 import { User } from "@blockframes/auth/+state/auth.store";
-import { ImgRef, createImgRef } from "@blockframes/utils/image-uploader";
+import { ImgRef, createImgRef } from "@blockframes/utils/media";
 import { PublicUser } from "./user.firestore";
 
 export * from './user.firestore';
