+++ libs/utils/src/lib/media/index.ts
@@ -1,0 +1,2 @@
+export * from './media.model';
+export * from './media.service';
