+++ libs/utils/src/lib/media/media.firestore.ts
@@ -1,0 +1,23 @@
+export interface ImgRef {
+  ref: string;
+  urls: {
+    original: string,
+    xs?: string,
+    md?: string,
+    lg?: string
+  };
+}
+
+export function createImgRef(ref: Partial<ImgRef> | string = {}): ImgRef {
+  const _ref = typeof ref === 'string' ? { urls: { original : ref } } : ref;
+  return {
+    ref: '',
+    urls: {
+      original: '',
+      xs: '',
+      md: '',
+      lg: ''
+    },
+    ..._ref
+  };
+}
