+++ libs/utils/src/lib/media/media.model.ts
@@ -1,0 +1,1 @@
+export * from './media.firestore';
